+++ app/src/main/java/kr/nbc/momo/presentation/mypage/MyPageFragment.kt
@@ -30,6 +30,7 @@ import kotlinx.coroutines.launch
 import kr.nbc.momo.R
 import kr.nbc.momo.databinding.FragmentMyPageBinding
 import kr.nbc.momo.presentation.UiState
+import kr.nbc.momo.presentation.group.read.EditReadGroupFragment
 import kr.nbc.momo.presentation.main.SharedViewModel
 import kr.nbc.momo.presentation.onboarding.GetStartedActivity
 import kr.nbc.momo.presentation.onboarding.signup.model.UserModel
@@ -47,47 +48,7 @@ class MyPageFragment : Fragment() {
     private val binding get() = _binding!!
     private val viewModel: MyPageViewModel by viewModels()
     private val sharedViewModel: SharedViewModel by activityViewModels()
-
-    //이미지 변경 상태 확인
-    private var isProfileImageChange = false
-    private var isBackgroundImageChange = false
-    private var isPortfolioImageChange = false
-
-    private var isEditMode = false
     private var currentUser: UserModel? = null
-    private var profileImageUri: Uri? = null
-    private var backgroundImageUri: Uri? = null
-    private var portfolioImageUri: Uri? = null
-
-    private val pickProfileImage = registerForActivityResult(ActivityResultContracts.PickVisualMedia()) { uri ->
-        if (uri != null) {
-            profileImageUri = uri
-            isProfileImageChange = true
-            binding.ivUserProfileImage.load(uri)
-        } else {
-            Log.d("PhotoPicker", "No media selected")
-        }
-    }
-
-    private val pickBackgroundImage = registerForActivityResult(ActivityResultContracts.PickVisualMedia()) { uri ->
-        if (uri != null) {
-            backgroundImageUri = uri
-            isBackgroundImageChange = true
-            binding.ivBackProfileThumbnail.load(uri)
-        } else {
-            Log.d("PhotoPicker", "No media selected")
-        }
-    }
-
-    private val pickPortfolioImage = registerForActivityResult(ActivityResultContracts.PickVisualMedia()) { uri ->
-        if (uri != null) {
-            portfolioImageUri = uri
-            isPortfolioImageChange = true
-            binding.ivPortfolioImage.load(uri)
-        } else {
-            Log.d("PhotoPicker", "No media selected")
-        }
-    }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
         _binding = FragmentMyPageBinding.inflate(inflater, container, false)
@@ -97,7 +58,7 @@ class MyPageFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         observeUserProfile()
-        eachEventHandler()
+        initEventHandler()
     }
 
     private fun observeUserProfile() {
@@ -133,6 +94,7 @@ class MyPageFragment : Fragment() {
             }
         }
     }
+
     private fun observeUserProfileUpdate() {
         viewLifecycleOwner.lifecycleScope.launch {
             repeatOnLifecycle(Lifecycle.State.STARTED) {
@@ -143,12 +105,14 @@ class MyPageFragment : Fragment() {
                             binding.includeUiState.setVisibleToVisible()
                             binding.scrollView.setVisibleToGone()
                         }
+
                         is UiState.Success -> {
                             Log.d("test", "Success")
                             binding.includeUiState.setVisibleToGone()
                             sharedViewModel.getCurrentUser()
                             binding.scrollView.setVisibleToVisible()
                         }
+
                         is UiState.Error -> {
                             Log.d("test", "Error")
                             binding.includeUiState.setVisibleToError()
@@ -167,10 +131,6 @@ class MyPageFragment : Fragment() {
             tvUserSelfIntroduction.text = user.userSelfIntroduction.ifEmpty { getString(R.string.introduce_yourself) }
             tvStackOfDevelopment.text = user.stackOfDevelopment.ifEmpty { getString(R.string.introduce_stack) }
             tvPortfolio.text = user.userPortfolioText.ifEmpty { getString(R.string.introduce_portfolio) }
-            etUserName.setText(user.userName)
-            etUserSelfIntroduction.setText(user.userSelfIntroduction)
-            etStackOfDevelopment.setText(user.stackOfDevelopment)
-            etPortfolio.setText(user.userPortfolioText)
             ivUserProfileImage.setThumbnailByUrlOrDefault(user.userProfileThumbnailUrl)
             ivBackProfileThumbnail.load(user.userBackgroundThumbnailUrl)
             ivPortfolioImage.setUploadImageByUrlOrDefault(user.userPortfolioImageUrl)
@@ -191,28 +151,14 @@ class MyPageFragment : Fragment() {
                 setSelectedChips(cgProgramTag, user.programOfDevelopment)
             }
         }
-
-        binding.etStackOfDevelopment.addTextWatcherWithError(500, "기술스택", binding.btnCompleteEdit, binding.tvCountStackEditText)
-        binding.etPortfolio.addTextWatcherWithError(500, "포트폴리오", binding.btnCompleteEdit, binding.tvCountPortfolioEditText)
-        binding.etUserSelfIntroduction.addTextWatcherWithError(60, "자기소개", binding.btnCompleteEdit)
-    }
-
-    private fun updateTextCount(et: EditText, tv: TextView) {
-        val textLength = et.text.length
-        val lengthText = "$textLength/500"
-        tv.text = lengthText
     }
 
     private fun clearUserInfo() {
         with(binding) {
             tvUserName.text = ""
-            etUserName.setText("")
             tvUserSelfIntroduction.text = ""
-            etUserSelfIntroduction.setText("")
             tvStackOfDevelopment.text = ""
-            etStackOfDevelopment.setText("")
             tvPortfolio.text = ""
-            etPortfolio.setText("")
             cgTypeTag.removeAllViews()
             cgProgramTag.removeAllViews()
             ivUserProfileImage.setThumbnailByUrlOrDefault(null)
@@ -221,154 +167,27 @@ class MyPageFragment : Fragment() {
         }
     }
 
-    private fun eachEventHandler() {
+    private fun initEventHandler() {
         binding.ivEditProfile.setOnClickListener {
-            setChangeMode()
-        }
-        binding.btnCompleteEdit.setOnClickListener {
-            if (validName()) {
-                saveProfileInfo()
-                setChangeMode()
-                requireActivity().hideKeyboard()
-                observeUserProfileUpdate()
-            }
-        }
-        binding.ivEditProfileImage.setOnClickListener {
-            pickProfileImage.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageAndVideo))
-        }
-        binding.ivEditBackProfileThumbnail.setOnClickListener {
-            pickBackgroundImage.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageAndVideo))
-        }
-        binding.btnGoOnBoarding.setOnClickListener {
-            val intent = Intent(requireActivity(), GetStartedActivity::class.java)
-            startActivity(intent)
-        }
-        binding.ivDeleteProfileImage.setOnClickListener {
-            profileImageUri = null
-            isProfileImageChange = true
-            binding.ivUserProfileImage.setThumbnailByUrlOrDefault(null)
-        }
-        binding.ivDeleteBackProfileThumbnail.setOnClickListener {
-            backgroundImageUri = null
-            isBackgroundImageChange = true
-            binding.ivBackProfileThumbnail.setImageResource(R.color.blue)
-        }
-        binding.ivDeletePortfolioImage.setOnClickListener {
-            portfolioImageUri = null
-            isPortfolioImageChange = true
-            binding.ivPortfolioImage.setUploadImageByUrlOrDefault(null)
+/*            val editMyPageFragment = EditMyPageFragment()
+            parentFragmentManager.beginTransaction()
+                .replace(R.id.fragment_container, editMyPageFragment)
+                .addToBackStack("MyPageFragment")
+                .commit()*/
         }
     }
 
-    private fun setChangeMode() {
-        isEditMode = !isEditMode
-
-        val chipGroupDev = resources.getStringArray(R.array.chipGroupDevelopmentOccupations)
-        val chipGroupLang = resources.getStringArray(R.array.chipProgramingLanguage)
-
-        if (isEditMode) {
-            setChipGroup(chipGroupDev, binding.cgTypeTag, isEditMode)
-            setChipGroup(chipGroupLang, binding.cgProgramTag, isEditMode)
-            selectChips(binding.cgTypeTag, currentUser?.typeOfDevelopment ?: emptyList())
-            selectChips(binding.cgProgramTag, currentUser?.programOfDevelopment ?: emptyList())
-            binding.tvEmptyTypeTag.setVisibleToGone()
-            binding.tvEmptyProgramTag.setVisibleToGone()
-            binding.cgTypeTag.setVisibleToVisible()
-            binding.cgProgramTag.setVisibleToVisible()
-            updateTextCount(binding.etStackOfDevelopment, binding.tvCountStackEditText)
-            updateTextCount(binding.etPortfolio, binding.tvCountPortfolioEditText)
-
-            binding.ivPortfolioImage.setOnClickListener {
-                pickPortfolioImage.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageAndVideo))
-            }
-            binding.llProfileImage.setBackgroundResource(R.drawable.circle_blue)
-            hideNav()
-            val rootView = requireActivity().window.decorView.findViewById<View>(android.R.id.content)
-            rootView.viewTreeObserver.addOnGlobalLayoutListener(globalLayoutListener)
-
-        } else {
-            //태그 빈값
-            if (binding.cgTypeTag.childCount == 0) {
-                binding.tvEmptyTypeTag.setVisibleToVisible()
-                binding.cgTypeTag.setVisibleToGone()
-            } else {
-                binding.tvEmptyTypeTag.setVisibleToGone()
-            }
-            if (binding.cgProgramTag.childCount == 0) {
-                binding.tvEmptyProgramTag.setVisibleToVisible()
-                binding.cgProgramTag.setVisibleToGone()
-            } else {
-                binding.tvEmptyProgramTag.setVisibleToGone()
-            }
-
-            setSelectedChips(binding.cgTypeTag, getChipText(binding.cgTypeTag))
-            setSelectedChips(binding.cgProgramTag, getChipText(binding.cgProgramTag))
-            currentUser?.let { initView(it) }
-
-            binding.ivPortfolioImage.setOnClickListener(null)
-            binding.llProfileImage.setBackgroundResource(0)
-            showNav()
-            val rootView = requireActivity().window.decorView.findViewById<View>(android.R.id.content)
-            rootView.viewTreeObserver.removeOnGlobalLayoutListener(globalLayoutListener)
-        }
-
-
-        val editMode = arrayOf(
-            binding.etUserName,
-            binding.etUserSelfIntroduction,
-            binding.etStackOfDevelopment,
-            binding.etPortfolio,
-            binding.btnCompleteEdit,
-            binding.ivEditProfileImage,
-            binding.ivEditBackProfileThumbnail,
-            binding.cvEditProfileImage,
-            binding.tvCountStackEditText,
-            binding.tvCountPortfolioEditText,
-            binding.cvDeleteProfileImage,
-            binding.ivDeleteBackProfileThumbnail,
-            binding.ivDeletePortfolioImage
-
-        )
-        val viewMode = arrayOf(
-            binding.ivEditProfile,
-            binding.tvUserName,
-            binding.tvUserSelfIntroduction,
-            binding.tvStackOfDevelopment,
-            binding.tvPortfolio,
-        )
-
-        editMode.forEach { if (isEditMode) it.setVisibleToVisible() else it.setVisibleToGone() }
-        viewMode.forEach { if (!isEditMode) it.setVisibleToVisible() else it.setVisibleToGone() }
-    }
-
-    private fun isLogin() { //todo 코드 간결화 필요
+    private fun isLogin() {
         binding.clUserDetailInfo.setVisibleToVisible()
         binding.ivEditProfile.setVisibleToVisible()
         binding.btnGoOnBoarding.setVisibleToGone()
     }
 
-    private fun isLogOut() { //todo 코드 간결화 필요
+    private fun isLogOut() {
         binding.clUserDetailInfo.setVisibleToGone()
         binding.ivEditProfile.setVisibleToGone()
         binding.tvUserName.setText(R.string.need_login)
         binding.btnGoOnBoarding.setVisibleToVisible()
-    }
-
-
-    private fun createChip(text: String, isEditMode: Boolean): Chip {
-        return Chip(requireContext()).apply {
-            this.text = text
-            this.isCheckable = true
-            this.isCloseIconVisible = false
-            this.isClickable = isEditMode
-            updateChipAppearance(this, isChecked)
-
-            setOnCheckedChangeListener { _, isChecked ->
-                if (isEditMode) {
-                    updateChipAppearance(this, isChecked)
-                }
-            }
-        }
     }
 
     private fun updateChipAppearance(chip: Chip, isChecked: Boolean) {
@@ -383,15 +202,7 @@ class MyPageFragment : Fragment() {
         )
     }
 
-    private fun setChipGroup(chipList: Array<String>, chipGroup: ChipGroup, isEditMode: Boolean) {
-        chipGroup.removeAllViews()
-        for (chipText in chipList) {
-            val chip = createChip(chipText, isEditMode)
-            chipGroup.addView(chip)
-        }
-    }
-
-    private fun setSelectedChips(chipGroup: ChipGroup, selectedChips: List<String>) { //선택된 chip만 보여줌
+    private fun setSelectedChips(chipGroup: ChipGroup, selectedChips: List<String>) {
         chipGroup.removeAllViews()
         selectedChips.forEach { chipText ->
             chipGroup.addView(Chip(requireContext()).apply {
@@ -401,98 +212,6 @@ class MyPageFragment : Fragment() {
                 updateChipAppearance(this, true)
             })
         }
-    }
-
-    private fun selectChips(chipGroup: ChipGroup, selectedChips: List<String>) {
-        chipGroup.children.forEach { chip ->
-            if (chip is Chip) {
-                chip.isChecked = selectedChips.contains(chip.text.toString())
-                updateChipAppearance(chip, chip.isChecked)
-                chip.isClickable = isEditMode
-            }
-        }
-    }
-
-    private fun saveProfileInfo() {
-        currentUser?.let { currentUser ->
-            val updatedUserModel = currentUser.copy(
-                userName = binding.etUserName.text.toString(),
-                userSelfIntroduction = binding.etUserSelfIntroduction.text.toString(),
-                stackOfDevelopment = binding.etStackOfDevelopment.text.toString(),
-                userPortfolioText = binding.etPortfolio.text.toString(),
-                typeOfDevelopment = getChipText(binding.cgTypeTag),
-                programOfDevelopment = getChipText(binding.cgProgramTag),
-                userProfileThumbnailUrl = if (isProfileImageChange) profileImageUri?.toString() ?: ""
-                else currentUser.userProfileThumbnailUrl,
-                userBackgroundThumbnailUrl = if (isBackgroundImageChange) backgroundImageUri?.toString() ?: ""
-                else currentUser.userBackgroundThumbnailUrl,
-                userPortfolioImageUrl = if (isPortfolioImageChange) portfolioImageUri?.toString() ?: ""
-                else currentUser.userPortfolioImageUrl
-            )
-            viewModel.saveUserProfile(updatedUserModel)
-            sharedViewModel.updateUser(updatedUserModel)
-            isProfileImageChange = false
-            isBackgroundImageChange = false
-            isPortfolioImageChange = false
-        }
-    }
-
-
-    private fun getChipText(chipGroup: ChipGroup): List<String> {
-        val textList = mutableListOf<String>()
-        for (i in 0 until chipGroup.childCount) {
-            val chip = chipGroup.getChildAt(i) as Chip
-            if (chip.isChecked) {
-                textList.add(chip.text.toString())
-            }
-        }
-        return textList
-    }
-
-
-    private fun validName(): Boolean {
-        val name = binding.etUserName.text.toString()
-        var isValid = true
-
-        if (name.isEmpty()) {
-            binding.etUserName.error = getString(R.string.please_edit_name)
-            isValid = false
-        } else if (!isValidName(name)) {
-            binding.etUserName.error = getString(R.string.edit_name_error)
-            isValid = false
-        } else {
-            binding.etUserName.error = null
-        }
-
-        return isValid
-    }
-
-    private fun isValidName(userName: String): Boolean {
-        val usernamePattern = "^[A-Za-z가-힣]{3,20}$"
-        return userName.matches(usernamePattern.toRegex())
-    }
-
-    private val globalLayoutListener = ViewTreeObserver.OnGlobalLayoutListener {
-        val rootView = requireActivity().window.decorView.findViewById<View>(android.R.id.content)
-        val rect = Rect()
-        rootView.getWindowVisibleDisplayFrame(rect)
-        val screenHeight = rootView.rootView.height
-        val keypadHeight = screenHeight - rect.bottom
-        if (keypadHeight > screenHeight * 0.15) {
-            binding.btnCompleteEdit.visibility = View.GONE
-        } else {
-            binding.btnCompleteEdit.visibility = View.VISIBLE
-        }
-    }
-
-    private fun showNav() {
-        val nav = requireActivity().findViewById<BottomNavigationView>(R.id.navigationView)
-        nav.setVisibleToVisible()
-    }
-
-    private fun hideNav() {
-        val nav = requireActivity().findViewById<BottomNavigationView>(R.id.navigationView)
-        nav.setVisibleToGone()
     }
 
     override fun onDestroyView() {
